[Fix] 거래후 회원 자산 관련 맞지 않는 부분 수정

See merge request s08-webmobile2-sub2/S08P12B111!185

+++ back/src/main/java/com/udteam/miristock/controller/MemberController.java
@@ -1,58 +1,75 @@
 package com.udteam.miristock.controller;
 
-import com.nimbusds.oauth2.sdk.ErrorResponse;
+import com.udteam.miristock.dto.MemberAdminDto;
 import com.udteam.miristock.dto.MemberDto;
 import com.udteam.miristock.service.MemberService;
 import com.udteam.miristock.util.HeaderUtil;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.Authorization;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.MessageSource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-
-import javax.servlet.http.HttpServletRequest;
-import java.util.List;
 
 @Slf4j
 @RestController
 @RequestMapping("/member")
 @RequiredArgsConstructor
+@Tag(name = "Member", description = "회원 정보")
 public class MemberController {
     private final MemberService memberservice;
 
-    @GetMapping
-    @ApiOperation(value ="모든 유저 호출 (TEST 용)")
-    public ResponseEntity<List<MemberDto>> selectAllMember(){
-        return ResponseEntity.ok().body(memberservice.selectAllMember());
-    }
-
     @DeleteMapping
-    @ApiOperation(value = "해당 유저 회원 탈퇴 처리")
+    @Operation(summary = "회원 탈퇴", description = "서비스에 가입된 회원 정보를 삭제한다.", tags = { "Member" })
     public ResponseEntity<String> deleteMember(@RequestHeader String Authorization){
+        log.debug("회원 삭제 요청됨");
+        MemberDto m = memberservice.selectOneMember(HeaderUtil.getAccessTokenString(Authorization));
+        log.info("회원 : {} | /member DELETE API호출 : 회원 탈퇴", m);
+        if(m.getMemberNo() == 1){
+            log.debug("관리자는 회원 탈퇴를 할 수 없습니다.");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("ADMIN Account Delete refusal");
+        }
         String userDetails = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         String token = HeaderUtil.getAccessTokenString(Authorization);
-        log.info("유저 디테일 = {}",userDetails);
-        if (memberservice.deleteMember(token)!= 0){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Error");
+        log.debug("유저 디테일 = {}",userDetails);
+        if (memberservice.deleteMember(token) != 0){
+            // 회원삭제 성공
+            return ResponseEntity.status(HttpStatus.OK).body("Delete Member Success");
         }
-        return ResponseEntity.ok().body("delete OK");
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Delete Member Fail");
     }
 
     @PutMapping("/nickname")
-    @ApiOperation(value = " 유저 닉네임 설정")
-    public ResponseEntity<MemberDto> updateMember(@RequestHeader String Authorization, @RequestParam String nickname){
-        String userDetails = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        String token = HeaderUtil.getAccessTokenString(Authorization);
-        log.info("유저 디테일 = {}",userDetails);
-        if(memberservice.selectOneMember(token)== null){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+    @Operation(summary = "회원 닉네임 수정", description = "서비스에 가입된 회원 닉네임 정보를 수정한다.", tags = { "Member" })
+    public ResponseEntity<?> updateMember(@RequestBody MemberDto memberDto){
+        log.info("회원 : {} | /member/nickname PUT API호출 : 회원 닉네임 수정", memberDto);
+        if(memberDto.getMemberNickname() == null || memberDto.getMemberNickname().equals("")){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("닉네임 Null 닉네임 수정 요청 거부");
+        } else if (memberDto.getMemberNickname().length() < 2 || memberDto.getMemberNickname().length() > 10){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("닉네임은 2자 이상 10자 이하만 입력할 수 있습니다.");
+        } else if (memberDto.getMemberNickname().equals("탈퇴유저")) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("설정할 수 없는 닉네임 입니다.");
+        } else if (memberDto.getMemberNickname().equals("관리자")) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("설정할 수 없는 닉네임 입니다.");
         }
-        return ResponseEntity.ok().body(memberservice.updateMember(token,nickname));
+        MemberDto result = memberservice.updateNickName(memberDto);
+        if(result == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("회원 정보 없음");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(result);
+    }
+
+
+    @PostMapping("/nicknamecheck")
+    @Operation(summary = "회원 닉네임 출력", description = "서비스에 가입된 회원 닉네임 정보를 출력한다.", tags = { "Member" })
+    public ResponseEntity<?> selectAllMember(@RequestBody MemberDto memberDto){
+        log.debug("유저 닉네임 체크 호출됨 : {} " , memberDto);
+        MemberAdminDto getMember = memberservice.selectOneMemberAllInfo(memberDto);
+        if(getMember == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("회원 정보를 찾을 수 없습니다.");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(getMember);
     }
 }
